✅ Commit: Admin protections, enriched user view, vehicle linking & webhook handling Changes included:
🔐 Only show admin sidebar links if user_metadata.role === 'admin'

✅ Show "Admin" badge in navbar if user is admin

👤 /admin/users now merges Supabase users with Enode users

Columns: user ID, name, email, admin yes/no, Enode linked yes/no, linked date

Added missing role and full_name to Supabase users table via auth trigger

🚗 Implemented /user/link-vehicle and /user/link-result

Validates linkToken and confirms link session for Enode

📡 Webhook system improvements:

Incoming /webhook/enode stores and processes vehicle events

save_vehicle_data_with_client now uses admin Supabase client

More detailed logging for vehicle and webhook actions

Added verify_signature() to app/enode/verify.py for future use

🧼 Moved all DB logic from route files into storage/ module

🔧 All Supabase interactions now use the correct client (user or admin)
