--- backend_bak/app/schemas/auth.py
@@ -1,19 +1,0 @@
-# backend/app/schemas/auth.py
-
-from pydantic import BaseModel, EmailStr
-
-class LoginRequest(BaseModel):
-    email: str
-    password: str
-
-class TokenResponse(BaseModel):
-    access_token: str
-    token_type: str = "bearer"
-
-class RegisterRequest(BaseModel):
-    name: str
-    email: EmailStr
-    password: str
-
-class RegisterResponse(BaseModel):
-    message: str
