--- backend_bak/app/storage/__init__.py
@@ -1,4 +1,0 @@
-# app/__init__.py
-
-# Required for Python to treat 'app' as a package.
-# No logic needed here for now.
