--- backend_bak/app/storage/db.py
@@ -1,10 +1,0 @@
-from dotenv import load_dotenv
-import os
-from supabase import create_client, Client
-
-load_dotenv()
-
-SUPABASE_URL = os.getenv("SUPABASE_URL")
-SUPABASE_KEY = os.getenv("SUPABASE_KEY")
-
-supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)
