--- backend_bak/pytest.ini
@@ -1,3 +1,0 @@
-# pytest.ini
-[pytest]
-pythonpath = .
