--- backend_bak/tests/test_internal_api.py
@@ -1,2 +1,0 @@
-def test_placeholder():
-    assert True
